split into directories
