+++ firmware/quintuna/BMS/src/app/segments/app_segments_broadcast.c
@@ -1,39 +1,15 @@
 #include "app_segments.h"
-#include "app_canTx.h"
+#include "app_segments_internal.h"
+#include "app_thermistor.h"
 #include "app_thermistors.h"
-#include "app_segments_setters.h"
-
-#include "app_utils.h"
-#include "io_log.h"
-#include "io_ltc6813.h"
-#include "io_time.h"
-
-#include <hw_utils.h>
-#include <math.h>
-#include <stdint.h>
-#include <stdlib.h>
+#include <app_canTx.h>
 #include <string.h>
 #include <float.h>
-
-// TODO: Take a closer look at these guys
-#define CONVERSION_TIME_MS (10)
-#define OWC_CONVERSION_TIME_MS (201)
-
-#define VUV (0x619U) // 2.5V Under-voltage threshold = (VUV + 1) * 16 * 100uV
-#define VOV (0xA41U) // 4.2V Over-voltage threshold = VOV * 16 * 100uV
-
-#define VREF_AUX_REG 5
-#define NUM_CONFIG_SYNC_TRIES 20
 
 #define CONVERT_100UV_TO_VOLTAGE(v_100uv) ((float)v_100uv * 1E-4f)
 #define CONVERT_VOLTAGE_TO_100UV(v) ((uint16_t)(v * 1E4f))
 
-typedef enum
-{
-    THERMISTOR_MUX_0_7,
-    THERMISTOR_MUX_8_13,
-    THERMISTOR_MUX_COUNT,
-} ThermistorMux;
+#define VREF_AUX_REG 5
 
 typedef struct
 {
@@ -41,300 +17,11 @@ typedef struct
     uint8_t cell;
 } CellPos;
 
-// Keeping separate buffers for every command is pretty wasteful but we have lots of RAM so oh well
-
-static SegmentConfig segment_config[NUM_SEGMENTS];
-static SegmentConfig read_segment_config[NUM_SEGMENTS];
-static ExitCode      config_success_buf[NUM_SEGMENTS];
-
-static uint16_t cell_voltage_regs[NUM_SEGMENTS][CELLS_PER_SEGMENT];
-static ExitCode cell_voltage_success[NUM_SEGMENTS][CELLS_PER_SEGMENT];
-static float    cell_voltages[NUM_SEGMENTS][CELLS_PER_SEGMENT];
-
-static uint16_t aux_regs[THERMISTOR_MUX_COUNT][NUM_SEGMENTS][AUX_REGS_PER_SEGMENT];
-static ExitCode aux_regs_success[THERMISTOR_MUX_COUNT][NUM_SEGMENTS][AUX_REGS_PER_SEGMENT];
-static float    segment_vref[NUM_SEGMENTS];
-static float    cell_temps[NUM_SEGMENTS][THERMISTORS_PER_SEGMENT];
-static bool     cell_temps_success[NUM_SEGMENTS][THERMISTORS_PER_SEGMENT];
-
-static StatusRegGroups status_regs[NUM_SEGMENTS];
-static ExitCode        status_regs_success[NUM_SEGMENTS];
-
-static uint16_t adc_accuracy_regs[NUM_SEGMENTS][CELLS_PER_SEGMENT];
-static ExitCode adc_accuracy_success[NUM_SEGMENTS][CELLS_PER_SEGMENT];
-
-static uint16_t voltage_self_test_regs[NUM_SEGMENTS][CELLS_PER_SEGMENT];
-static ExitCode voltage_self_test_success[NUM_SEGMENTS][CELLS_PER_SEGMENT];
-
-static uint16_t aux_self_test_regs[NUM_SEGMENTS][AUX_REGS_PER_SEGMENT];
-static ExitCode aux_self_test_reg_success[NUM_SEGMENTS][AUX_REGS_PER_SEGMENT];
-
-static StatusRegGroups status_self_test_regs[NUM_SEGMENTS];
-static ExitCode        status_self_test_success[NUM_SEGMENTS];
-
-static uint16_t owc_pucv_regs[NUM_SEGMENTS][CELLS_PER_SEGMENT];
-static ExitCode owc_pucv_success[NUM_SEGMENTS][CELLS_PER_SEGMENT];
-static uint16_t owc_pdcv_regs[NUM_SEGMENTS][CELLS_PER_SEGMENT];
-static ExitCode owc_pdcv_success[NUM_SEGMENTS][CELLS_PER_SEGMENT];
-static bool     owc_results[NUM_SEGMENTS][CELL_TAPS_PER_SEGMENT];
-static bool     owc_results_success[NUM_SEGMENTS][CELL_TAPS_PER_SEGMENT];
-
-static float pack_voltage;
-static float max_cell_voltage;
-static float min_cell_voltage;
-static float max_cell_temp;
-static float min_cell_temp;
-
-static void writeThermistorMux(ThermistorMux mux)
-{
-    for (uint8_t seg = 0; seg < NUM_SEGMENTS; seg++)
-    {
-        SegmentConfig *const out       = &segment_config[seg];
-        CFGBR *const         seg_b_cfg = &out->reg_b;
-
-        if (mux == THERMISTOR_MUX_0_7)
-        {
-            seg_b_cfg->gpio_6_9 = 0x7;
-        }
-        else
-        {
-            seg_b_cfg->gpio_6_9 = 0xF;
-        }
-    }
-}
-
-void app_segments_writeDefaultConfig(void)
-{
-    for (uint8_t seg = 0; seg < NUM_SEGMENTS; seg++)
-    {
-        SegmentConfig *const out       = &segment_config[seg];
-        CFGAR *const         seg_a_cfg = &out->reg_a;
-        CFGBR *const         seg_b_cfg = &out->reg_b;
-
-        // enables GPIOs for thermistors (reading values)
-        seg_a_cfg->gpio_1_5 = 0x1F;
-        seg_b_cfg->gpio_6_9 = 0xF;
-        seg_a_cfg->refon    = 1;
-        seg_a_cfg->dten     = 0;
-
-        // upper and lower bound
-        seg_a_cfg->vuv_0_7  = VUV & 0xFF;
-        seg_a_cfg->vuv_8_11 = VUV >> 8 & 0xF;
-        seg_a_cfg->vov_0_3  = VOV & 0xF;
-        seg_a_cfg->vov_4_11 = VOV >> 4 & 0xFF;
-    }
-}
-
-void app_segments_setBalanceConfig(const bool balance_config[NUM_SEGMENTS][CELLS_PER_SEGMENT])
-{
-    for (uint8_t seg = 0; seg < NUM_SEGMENTS; seg++)
-    {
-        SegmentConfig *const out       = &segment_config[seg];
-        CFGAR *const         seg_a_cfg = &out->reg_a;
-        CFGBR *const         seg_b_cfg = &out->reg_b;
-        uint32_t             dcc_bits  = 0U;
-
-        // Get dcc bits to write for the current segment (which cells to balance)
-        for (uint8_t cell = 0; cell < CELLS_PER_SEGMENT; cell++)
-        {
-            dcc_bits |= (uint32_t)(balance_config[seg][cell] << cell);
-        }
-
-        // TODO double check this is correct
-        // note that we want balance_config to be auto scalable (for futureproofing)
-        seg_b_cfg->dcc_0     = 0;
-        seg_a_cfg->dcc_1_8   = dcc_bits & 0xFF;
-        seg_a_cfg->dcc_9_12  = dcc_bits >> 8 & 0xF;
-        seg_b_cfg->dcc_13_16 = dcc_bits >> 12 & 0xF;
-        seg_b_cfg->dcc_18    = 0;
-        seg_b_cfg->dcc_17    = 0;
-    }
-}
-
-/**
- * @returns whether there are no errors and config matches
- * Hence if false there are either errors or the config does not match
- */
-static bool isConfigEqual(void)
-{
-    for (uint8_t try = 0; try < 3; try++)
-    {
-        io_ltc6813_readConfigurationRegisters(read_segment_config, config_success_buf);
-
-        bool all_ok = true;
-        for (uint8_t seg = 0; seg < NUM_SEGMENTS; seg++)
-        {
-            if (IS_EXIT_ERR(config_success_buf[seg]))
-            {
-                all_ok = false;
-            }
-        }
-
-        if (all_ok)
-        {
-            break;
-        }
-    }
-
-    bool seg_config_same = true;
-    for (uint8_t seg = 0; seg_config_same && seg < NUM_SEGMENTS; seg++)
-    {
-        const SegmentConfig *const a = &segment_config[seg];
-        const SegmentConfig *const b = &read_segment_config[seg];
-
-        // NOTE: the reason we don't use memcmp is because we only really care about these configs
-        // base configs
-        seg_config_same &= a->reg_a.adcopt == b->reg_a.adcopt;
-        seg_config_same &= a->reg_a.refon == b->reg_a.refon;
-
-        // undervoltage
-        seg_config_same &= a->reg_a.vuv_0_7 == b->reg_a.vuv_0_7;
-        seg_config_same &= a->reg_a.vuv_8_11 == b->reg_a.vuv_8_11;
-
-        // overvoltage config
-        seg_config_same &= a->reg_a.vov_0_3 == b->reg_a.vov_0_3;
-        seg_config_same &= a->reg_a.vov_4_11 == b->reg_a.vov_4_11;
-
-        // balancing leds
-        seg_config_same &= a->reg_b.dcc_0 == b->reg_b.dcc_0;
-        seg_config_same &= a->reg_a.dcc_1_8 == b->reg_a.dcc_1_8;
-        seg_config_same &= a->reg_a.dcc_9_12 == b->reg_a.dcc_9_12;
-        seg_config_same &= a->reg_b.dcc_13_16 == b->reg_b.dcc_13_16;
-        seg_config_same &= a->reg_b.dcc_17 == b->reg_b.dcc_17;
-        seg_config_same &= a->reg_b.dcc_18 == b->reg_b.dcc_18;
-    }
-
-    return seg_config_same;
-}
-
-ExitCode app_segments_configSync(void)
-{
-    for (int tries = 0; tries < NUM_CONFIG_SYNC_TRIES; tries++)
-    {
-        const bool is_config_equal = isConfigEqual();
-
-        // first check if it's a comms thing
-        bool comms_ok = true;
-        for (uint8_t seg = 0; seg < NUM_SEGMENTS; seg++)
-        {
-            const bool seg_comms_ok = IS_EXIT_OK(config_success_buf[seg]);
-            comm_ok_setters[seg](seg_comms_ok);
-            comms_ok &= seg_comms_ok;
-        }
-
-        if (is_config_equal)
-        {
-            return EXIT_CODE_OK;
-        }
-
-        if (comms_ok)
-        {
-            // comms is fine, but configs are not matching
-            LOG_IF_ERR(io_ltc6813_writeConfigurationRegisters(segment_config));
-        }
-        else
-        {
-            // comms are bad, reset voltages and temps in the CAN table
-            // TODO add a function to reset all values??
-            for (uint8_t seg_idx = 0; seg_idx < NUM_SEGMENTS; seg_idx++)
-            {
-                for (uint8_t cell = 0; cell < CELLS_PER_SEGMENT; cell++)
-                {
-                    cell_voltage_setters[seg_idx][cell](-0.1f);
-                }
-
-                for (uint8_t thermistor = 0; thermistor < CELLS_PER_SEGMENT; thermistor++)
-                {
-                    thermistor_setters[seg_idx][thermistor](-1);
-                }
-            }
-        }
-    }
-
-    return EXIT_CODE_TIMEOUT;
-}
-
-ExitCode app_segments_runVoltageConversion(void)
-{
-    RETURN_IF_ERR(io_ltc6813_startCellsAdcConversion());
-    io_time_delay(CONVERSION_TIME_MS);
-    io_ltc6813_readVoltageRegisters(cell_voltage_regs, cell_voltage_success);
-    return EXIT_CODE_OK;
-}
-
-ExitCode app_segments_runAuxConversion(void)
-{
-    for (ThermistorMux mux = 0U; mux < THERMISTOR_MUX_COUNT; mux++)
-    {
-        writeThermistorMux(mux);
-        io_ltc6813_writeConfigurationRegisters(segment_config);
-
-        RETURN_IF_ERR(io_ltc6813_startThermistorsAdcConversion());
-        io_time_delay(CONVERSION_TIME_MS);
-        io_ltc6813_readAuxRegisters(aux_regs[mux], aux_regs_success[mux]);
-    }
-
-    return EXIT_CODE_OK;
-}
-
-ExitCode app_segments_runStatusConversion(void)
-{
-    RETURN_IF_ERR(io_ltc6813_diagnoseMUX());
-    io_time_delay(1); // Empirically you need this small delay after sending DIAGN
-    RETURN_IF_ERR(io_ltc6813_startInternalADCConversions());
-    io_time_delay(CONVERSION_TIME_MS);
-    io_ltc6813_getStatus(status_regs, status_regs_success);
-    return EXIT_CODE_OK;
-}
-
-ExitCode app_segments_runAdcAccuracyTest(void)
-{
-    RETURN_IF_ERR(io_ltc6813_overlapADCTest());
-    io_time_delay(CONVERSION_TIME_MS);
-    io_ltc6813_readVoltageRegisters(adc_accuracy_regs, adc_accuracy_success);
-    return EXIT_CODE_OK;
-}
-
-ExitCode app_segments_runVoltageSelfTest(void)
-{
-    RETURN_IF_ERR(io_ltc6813_sendSelfTestVoltages());
-    io_time_delay(CONVERSION_TIME_MS);
-    io_ltc6813_readVoltageRegisters(voltage_self_test_regs, voltage_self_test_success);
-    return EXIT_CODE_OK;
-}
-
-ExitCode app_segments_runAuxSelfTest(void)
-{
-    RETURN_IF_ERR(io_ltc6813_sendSelfTestAux());
-    io_time_delay(CONVERSION_TIME_MS);
-    io_ltc6813_readAuxRegisters(aux_self_test_regs, aux_self_test_reg_success);
-    return EXIT_CODE_OK;
-}
-
-ExitCode app_segments_runStatusSelfTest(void)
-{
-    RETURN_IF_ERR(io_ltc6813_sendSelfTestStat());
-    io_time_delay(CONVERSION_TIME_MS);
-    io_ltc6813_getStatus(status_self_test_regs, status_self_test_success);
-    return EXIT_CODE_OK;
-}
-
-ExitCode app_segments_runOpenWireCheck(void)
-{
-    RETURN_IF_ERR(io_ltc6813_owcPull(PULL_UP));
-    io_time_delay(OWC_CONVERSION_TIME_MS);
-    RETURN_IF_ERR(io_ltc6813_owcPull(PULL_UP));
-    io_time_delay(OWC_CONVERSION_TIME_MS);
-    io_ltc6813_readVoltageRegisters(owc_pucv_regs, owc_pucv_success);
-
-    RETURN_IF_ERR(io_ltc6813_owcPull(PULL_DOWN));
-    io_time_delay(OWC_CONVERSION_TIME_MS);
-    RETURN_IF_ERR(io_ltc6813_owcPull(PULL_DOWN));
-    io_time_delay(OWC_CONVERSION_TIME_MS);
-    io_ltc6813_readVoltageRegisters(owc_pdcv_regs, owc_pdcv_success);
-
-    return EXIT_CODE_OK;
-}
+float pack_voltage;
+float max_cell_voltage;
+float min_cell_voltage;
+float max_cell_temp;
+float min_cell_temp;
 
 void app_segments_broadcastCellVoltages(void)
 {
@@ -672,29 +359,4 @@ void app_segments_broadcastTempStats(void)
     app_canTx_BMS_MinCellTempIdx_set(min_temp_pos.cell);
     app_canTx_BMS_MinCellTemp_set(tmp_min_cell_temp);
     app_canTx_BMS_AvgCellTemp_set(temp_sum / (NUM_SEGMENTS * CELLS_PER_SEGMENT));
-}
-
-float app_segments_getPackVoltage()
-{
-    return pack_voltage;
-}
-
-float app_segments_getMaxCellVoltage()
-{
-    return max_cell_voltage;
-}
-
-float app_segments_getMinCellVoltage()
-{
-    return min_cell_voltage;
-}
-
-float app_segments_getMaxCellTemp()
-{
-    return max_cell_temp;
-}
-
-float app_segments_getMinCellTemp()
-{
-    return min_cell_temp;
 }
